+++ grid2op/Action/PowerlineChangeAction.py
@@ -17,13 +17,14 @@ class PowerlineChangeAction(PlayableAction):
     Nothing else is supported and any attempt to use something else 
     will have no impact.
     """
-    def __init__(self, gridobj):
-        super().__init__(gridobj)
+    authorized_keys = {
+        "change_line_status"
+    }
 
-        self.authorized_keys = {
-            "change_line_status"
-        }
+    attr_list_vect = [
+        "_switch_line_status"
+    ]
+    attr_list_set = set(attr_list_vect)
 
-        self.attr_list_vect = [
-            "_switch_line_status"
-        ]
+    def __init__(self):
+        super().__init__()