+++ CHANGELOG.rst
@@ -30,14 +30,18 @@ Change Log
   the same idea: impossibility to reconnect a powerilne. **Side effect** observation have a different size now (
   merging of `obs.time_before_line_reconnectable` into `obs.time_before_cooldown_line`). Size is now reduce of
   the number of powerlines of the grid.
+- [BREAKING]  merge of `act.vars_action` into `env.attr_list_vect` which implemented the same concepts.
 - [FIXED] `grid2op.PlotGrid` rounding error when casting from np.float32 to python.float
 - [FIXED] `grid2op.BaseEnv.fast_forward_chronics` Calls the correct methods and is now working properly
+- [FIXED] `__iadd__` is now properly implemented for the action with proper care given to action types.
 - [UPDATED] Notebooks are updated to reflect API improvements changes
 - [UPDATED] `grid2op.make` can now handle the download & caching of datasets
 - [UPDATED] Test/Sample datasets provide datetime related files .info
 - [UPDATED] Test/Sample datasets grid_layout.json
 - [UPDATED] `grid2op.PlotGrid` Color schemes and optional infos displaying
 - [UPDATED] `grid2op.Episode.EpisodeReplay` Improved gif output performance
+- [UPDATED] Action and Observation are now created without having to call `init_grid(gridobject)` which lead to
+  small speed up and memory saving.
 
 [0.7.1] - 2020-04-22
 ----------------------